requirements.txt
ultralytics==8.3.204
python-dotenv==1.1.1